requirements.txt
alembic==1.17.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
Pillow==10.0.1
psycopg==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
SQLAlchemy==2.0.43
tomli==2.3.0
tools==1.0.21
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uWSGI==2.0.30
Werkzeug==2.3.7
WTForms==3.0.1
